requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
decorator==5.1.1
Deprecated==1.2.14
discord.py==2.3.2
fabric2==3.2.2
frozenlist==1.4.0
idna==3.4
invoke==2.2.0
multidict==6.0.4
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.6
wrapt==1.15.0
yarl==1.9.2
